Provide missing endpoints, replace decimal with float fields

+++ src/backend/actions/ci_get_cell_data.py
@@ -31,8 +31,8 @@ class CellDataResponseSchema(Schema):
     etypes = fields.List(cls_or_instance=fields.String())
     morphology_classes = fields.List(cls_or_instance=fields.String())
     layers = fields.List(cls_or_instance=fields.String())
-    positions = fields.List(cls_or_instance=fields.Decimal())
-    orientations = fields.List(cls_or_instance=fields.Decimal())
+    positions = fields.List(cls_or_instance=fields.Float())
+    orientations = fields.List(cls_or_instance=fields.Float())
 
 
 class CIGetCellData(Action):
